Introduced enum class GameState to replace defines
While not strictly the same because I changed the values and made GameState equivalent to an unsigned char, this approach has several advantages:
- Implicit type-safety
- The actual values behind the names no longer matter to the programmer
- Any additional states can be added without having to designate values to them
- Added meaning to values (`return 1;` is no longer valid, the value is forced to be unambiguous)

Some of the space saving is from using `unsigned char` as the backing type (`unsigned char` actually results in less space usage than `char` because `char` is signed and signed numbers involve more work even when you aren't doing any arithmetic).
The rest of the space saving is the fact the values are now linear instead of jumping from 2 to 247. This allows the compiler to optimise the jumps used in the switch statement.